Moved java classes to new package
