--- src/main/java/uk.gov.hmcts.reform.fpl.ccddatamigration/domain/common/EmailAddress.java
@@ -1,13 +1,0 @@
-package uk.gov.hmcts.reform.fpl.ccddatamigration.domain.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class EmailAddress {
-    private final String email;
-    private final String emailUsageType;
-}
