--- src/main/java/uk.gov.hmcts.reform.fpl.ccddatamigration/domain/common/Party.java
@@ -1,43 +1,0 @@
-package uk.gov.hmcts.reform.fpl.ccddatamigration.domain.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class Party {
-    private final String partyID;
-    private final String idamID;
-    private final String partyType;
-    private final String title;
-    private final String firstName;
-    private final String lastName;
-    private final String organisationName;
-    private final String dateOfBirth;
-    private final Address address;
-    private final EmailAddress email;
-    private final String gender;
-    private final String genderIdentification;
-    private final String livingSituation;
-    private final String situationDetails;
-    private final String situationDate;
-    private final String keyDates;
-    private final String careAndContact;
-    private final String adoption;
-    private final String placementOrderApplication;
-    private final String placementCourt;
-    private final String mothersName;
-    private final String fathersName;
-    private final String fathersResponsibility;
-    private final String socialWorkerName;
-    private final String socialWorkerTel;
-    private final String additionalNeeds;
-    private final String additionalNeedsDetails;
-    private final String detailsHidden;
-    private final String detailsHiddenReason;
-    private final String litigationIssues;
-    private final String litigationIssuesDetails;
-    private final TelephoneNumber telephoneNumber;
-}
