--- src/main/java/uk.gov.hmcts.reform.fpl.ccddatamigration/domain/common/TelephoneNumber.java
@@ -1,14 +1,0 @@
-package uk.gov.hmcts.reform.fpl.ccddatamigration.domain.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class TelephoneNumber {
-    private final String telephoneNumber;
-    private final String telephoneUsageType;
-    private final String contactDirection;
-}
