+++ src/main/java/uk/gov/hmcts/reform/domain/Child.java
@@ -1,9 +1,9 @@
-package uk.gov.hmcts.reform.fpl.ccddatamigration.domain;
+package uk.gov.hmcts.reform.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import uk.gov.hmcts.reform.fpl.ccddatamigration.domain.common.Party;
+import uk.gov.hmcts.reform.domain.common.Party;
 
 @Data
 @Builder
