+++ src/main/java/uk/gov/hmcts/reform/domain/OldChild.java
@@ -1,13 +1,11 @@
-package uk.gov.hmcts.reform.fpl.ccddatamigration.domain;
+package uk.gov.hmcts.reform.domain;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import uk.gov.hmcts.reform.fpl.ccddatamigration.domain.common.Address;
 import lombok.Builder;
 import lombok.Data;
-
-import java.util.Date;
+import uk.gov.hmcts.reform.domain.common.Address;
 
 @Data
 @Builder
