+++ src/main/resources/application.properties
@@ -14,7 +14,7 @@ idam.s2s-auth.url=http://localhost:4502
 ccd.jurisdiction=PUBLICLAW
 ccd.casetype=CARE_SUPERVISION_EPO
 
-ccd.caseId=
+ccd.caseId=1561721008740557
 
 core_case_data.api.url=http://localhost:4452
